removed quality select through context menu entry before playback
Calling the same Plugin through the context menu of an item
does not work as expected. As a result the video either won't play
or (with some workarounds) will play but all information like
title and so on is lost (in essence there will be no link between
the video played and the item of the video in the list view -
so no information about how much time I have spent watching
this stream and so on). Since I value these informations I
removed the option to start playback of streams via the context
menu with selecting the quality before hand. Instead I have added
a new quality option in the settings of the addon - "always ask",
which will always ask in which quality to play a stream before
the actual playback.