fixed playback of lower resolutions
This was introduced with the new stream qualities
used on the TI6 streams. Since we won't play them
back anyway and we just need the qualities in our
list of qualities so we won't get an error I moved
them to the end of list. Now the index of the
stream quality that can be selected in the addon
corresponds to the index of the quality in the
quality list again.