added new quality 1080p60 to the sorting list
Twitch decided to not call the 'Source' quality in
the TI6 streams 'Source' anymore. Instead it's
called 1080p60 now. I added this to the quality sort
list. Also I made the sort more robust concerning
unknown qualities. It is now assumed they are worse
than anything else and are thus added to the end of
the sorted quality list.